+++ frontend/common/utils/utils.tsx
@@ -662,17 +662,6 @@ const Utils = Object.assign({}, require('./base/_utils'), {
     return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')
   },
 
-  openChat() {
-    const usePylon = flagsmith.hasFeature('pylon_chat')
-
-    if (usePylon && typeof window.Pylon !== 'undefined') {
-      window.Pylon('show')
-      Utils.setupPylon()
-    } else if (typeof $crisp !== 'undefined') {
-      $crisp.push(['do', 'chat:open'])
-      Utils.setupCrisp()
-    }
-  },
   removeElementFromArray(array: any[], index: number) {
     return array.slice(0, index).concat(array.slice(index + 1))
   },
@@ -693,84 +682,6 @@ const Utils = Object.assign({}, require('./base/_utils'), {
     return `${value}`
   },
 
-  setupCrisp() {
-    const user = AccountStore.model as AccountModel
-    if (typeof $crisp === 'undefined' || !user) {
-      return
-    }
-    $crisp.push([
-      'set',
-      'session:data',
-      [[['hosting', Utils.isSaas() ? 'SaaS' : 'Self-Hosted']]],
-    ])
-    const organisation = AccountStore.getOrganisation() as Organisation
-    const formatOrganisation = (o: Organisation) => {
-      const plan = AccountStore.getActiveOrgPlan()
-      return `${o.name} (${plan}) #${o.id}`
-    }
-    const otherOrgs = user?.organisations.filter(
-      (v) => v.id !== organisation?.id,
-    )
-    if (window.$crisp) {
-      $crisp.push(['set', 'user:email', user.email])
-      $crisp.push(['set', 'user:nickname', `${getUserDisplayName(user)}`])
-      if (otherOrgs.length) {
-        $crisp.push([
-          'set',
-          'session:data',
-          [[['other-orgs', `${otherOrgs?.length} other organisations`]]],
-        ])
-      }
-      $crisp.push([
-        'set',
-        'session:data',
-        [
-          [
-            ['user-id', `${user.id}`],
-            [
-              'date-joined',
-              `${moment(user.date_joined).format('Do MMM YYYY')}`,
-            ],
-          ],
-        ],
-      ])
-      if (organisation) {
-        $crisp.push(['set', 'user:company', formatOrganisation(organisation)])
-        $crisp.push([
-          'set',
-          'session:data',
-          [[['seats', organisation.num_seats]]],
-        ])
-      }
-    }
-  },
-  setupPylon() {
-    const user = AccountStore.model as AccountModel
-    if (typeof window.Pylon === 'undefined' || !user) {
-      return
-    }
-    const organisation = AccountStore.getOrganisation() as Organisation
-    const plan = AccountStore.getActiveOrgPlan()
-    const otherOrgs = user?.organisations.filter(
-      (v) => v.id !== organisation?.id,
-    )
-
-    try {
-      // Set Pylon configuration using the documented window.pylon object
-      window.pylon = {
-        chat_settings: {
-          account_id: String(user.id),
-          app_id: Project.pylonAppId,
-          email: user.email,
-          name: getUserDisplayName(user),
-          // Store additional metadata as custom fields
-          // Note: Pylon may support custom fields - check their docs for metadata capabilities
-        },
-      }
-    } catch (error) {
-      console.error('Error setting up Pylon:', error)
-    }
-  },
   tagDisabled: (tag: Tag | undefined) => {
     const hasStaleFlagsPermission = Utils.getPlansPermission('STALE_FLAGS')
     return tag?.type === 'STALE' && !hasStaleFlagsPermission
